chore: fix alembic nonsense
