+++ alembic/env.py
@@ -4,7 +4,7 @@ from logging.config import fileConfig
 from sqlalchemy import engine_from_config, pool
 
 from alembic import context
-from letta.metadata import Base
+from letta.base import Base
 from letta.settings import settings
 
 # this is the Alembic Config object, which provides
@@ -72,7 +72,7 @@ def run_migrations_online() -> None:
     )
 
     with connectable.connect() as connection:
-        context.configure(connection=connection, target_metadata=target_metadata)
+        context.configure(connection=connection, target_metadata=target_metadata, include_schemas=True)
 
         with context.begin_transaction():
             context.run_migrations()
