--- alembic/versions/34836dc9e903_init_again.py
@@ -1,147 +1,0 @@
-"""init-again
-
-Revision ID: 34836dc9e903
-Revises:
-Create Date: 2024-10-11 10:46:24.779701
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-
-from alembic import op
-from letta.metadata import EmbeddingConfigColumn, LLMConfigColumn
-
-# revision identifiers, used by Alembic.
-revision: str = "34836dc9e903"
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "agent_source_mapping",
-        sa.Column("id", sa.String(), nullable=False),
-        sa.Column("user_id", sa.String(), nullable=False),
-        sa.Column("agent_id", sa.String(), nullable=False),
-        sa.Column("source_id", sa.String(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index("agent_source_mapping_idx_user", "agent_source_mapping", ["user_id", "agent_id", "source_id"], unique=False)
-    op.create_table(
-        "agents",
-        sa.Column("id", sa.String(), nullable=False),
-        sa.Column("user_id", sa.String(), nullable=False),
-        sa.Column("name", sa.String(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), server_default=sa.text("now()"), nullable=True),
-        sa.Column("description", sa.String(), nullable=True),
-        sa.Column("message_ids", sa.JSON(), nullable=True),
-        sa.Column("memory", sa.JSON(), nullable=True),
-        sa.Column("system", sa.String(), nullable=True),
-        sa.Column("tools", sa.JSON(), nullable=True),
-        sa.Column("agent_type", sa.String(), nullable=True),
-        sa.Column("llm_config", LLMConfigColumn(), nullable=True),
-        sa.Column("embedding_config", EmbeddingConfigColumn(), nullable=True),
-        sa.Column("metadata_", sa.JSON(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index("agents_idx_user", "agents", ["user_id"], unique=False)
-    op.create_table(
-        "block",
-        sa.Column("id", sa.String(), nullable=False),
-        sa.Column("value", sa.String(), nullable=False),
-        sa.Column("limit", sa.BIGINT(), nullable=True),
-        sa.Column("name", sa.String(), nullable=False),
-        sa.Column("template", sa.Boolean(), nullable=True),
-        sa.Column("label", sa.String(), nullable=True),
-        sa.Column("metadata_", sa.JSON(), nullable=True),
-        sa.Column("description", sa.String(), nullable=True),
-        sa.Column("user_id", sa.String(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index("block_idx_user", "block", ["user_id"], unique=False)
-    op.create_table(
-        "jobs",
-        sa.Column("id", sa.String(), nullable=False),
-        sa.Column("user_id", sa.String(), nullable=True),
-        sa.Column("status", sa.String(), nullable=True),
-        sa.Column("created_at", sa.DateTime(timezone=True), server_default=sa.text("now()"), nullable=True),
-        sa.Column("completed_at", sa.DateTime(timezone=True), nullable=True),
-        sa.Column("metadata_", sa.JSON(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "organizations",
-        sa.Column("id", sa.String(), nullable=False),
-        sa.Column("name", sa.String(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "sources",
-        sa.Column("id", sa.String(), nullable=False),
-        sa.Column("user_id", sa.String(), nullable=False),
-        sa.Column("name", sa.String(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), server_default=sa.text("now()"), nullable=True),
-        sa.Column("embedding_config", EmbeddingConfigColumn(), nullable=True),
-        sa.Column("description", sa.String(), nullable=True),
-        sa.Column("metadata_", sa.JSON(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index("sources_idx_user", "sources", ["user_id"], unique=False)
-    op.create_table(
-        "tokens",
-        sa.Column("id", sa.String(), nullable=False),
-        sa.Column("user_id", sa.String(), nullable=False),
-        sa.Column("key", sa.String(), nullable=False),
-        sa.Column("name", sa.String(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index("tokens_idx_key", "tokens", ["key"], unique=False)
-    op.create_index("tokens_idx_user", "tokens", ["user_id"], unique=False)
-    op.create_table(
-        "tools",
-        sa.Column("id", sa.String(), nullable=False),
-        sa.Column("name", sa.String(), nullable=False),
-        sa.Column("user_id", sa.String(), nullable=True),
-        sa.Column("description", sa.String(), nullable=True),
-        sa.Column("source_type", sa.String(), nullable=True),
-        sa.Column("source_code", sa.String(), nullable=True),
-        sa.Column("json_schema", sa.JSON(), nullable=True),
-        sa.Column("module", sa.String(), nullable=True),
-        sa.Column("tags", sa.JSON(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "users",
-        sa.Column("id", sa.String(), nullable=False),
-        sa.Column("org_id", sa.String(), nullable=True),
-        sa.Column("name", sa.String(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=True),
-        sa.Column("policies_accepted", sa.Boolean(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("users")
-    op.drop_table("tools")
-    op.drop_index("tokens_idx_user", table_name="tokens")
-    op.drop_index("tokens_idx_key", table_name="tokens")
-    op.drop_table("tokens")
-    op.drop_index("sources_idx_user", table_name="sources")
-    op.drop_table("sources")
-    op.drop_table("organizations")
-    op.drop_table("jobs")
-    op.drop_index("block_idx_user", table_name="block")
-    op.drop_table("block")
-    op.drop_index("agents_idx_user", table_name="agents")
-    op.drop_table("agents")
-    op.drop_index("agent_source_mapping_idx_user", table_name="agent_source_mapping")
-    op.drop_table("agent_source_mapping")
-    # ### end Alembic commands ###
